Refactoring: new controllers
Missions are matched to their head of mission then they are triaged into different paths.
