+++ controllers/user_controllers.py
@@ -1,7 +1,7 @@
 from fastapi import HTTPException, status
 
 
-from models.user_model import UserIn
+from schemas.user_models import UserIn
 from mongodb.user import UserBase
 from auth.password_hasher import get_password_hash
