--- models/message_models.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel
-from typing import Optional
-
-
-class Message(BaseModel):
-    message: Optional[str] | None = "Sample message"
