+++ routes/embassy_routes.py
@@ -4,7 +4,7 @@ Embassy router
 
 from fastapi import APIRouter
 from fastapi.security import OAuth2PasswordBearer
-from controllers import in_collection, from_missions
+from controllers import from_missions
 from mongodb.mission import MissionUnion
 
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
@@ -12,17 +12,11 @@ oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
 router = APIRouter()
 
 
-# Create
-
-
 @router.get("/")
 async def all_embassies():
-    # return await in_collection.embassies_query_all()
-    # return await in_collection.embassies_return_all()
-
     return await from_missions.get_embassies()
 
 
-@router.get("/{id}")
+@router.get("/{embassy_id}")
 async def by_id(embassy_id: str):
     return await from_missions.get_mission_by_id(embassy_id)
